+++ src/assets/rest/rest.py
@@ -116,10 +116,10 @@ def delayed_vpn_server_change():
 # region
 
 
-@app.get("/get_interfaces1")
-async def get_interfaces1():
+@app.get("/get_interfaces")
+async def get_interfaces():
     try:
-        # test: http://localhost:8080/get_interfaces1
+        # test: http://localhost:8080/get_interfaces
         debug = False
         ret_obj = {}